* fixed some unsigned overflows
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@1383 29311d96-e01e-0410-9327-a35deaab8ce9

+++ trunk/ceph/mds/Server.cc
@@ -723,7 +723,7 @@ CInode* Server::rdlock_path_pin_ref(MDRequest *mdr, bool want_auth)
   // lock the path
   set<SimpleLock*> rdlocks, empty;
 
-  for (unsigned i=0; i<trace.size(); i++) 
+  for (int i=0; i<(int)trace.size(); i++) 
     rdlocks.insert(&trace[i]->lock);
 
   if (!mds->locker->acquire_locks(mdr, rdlocks, empty, empty))
@@ -797,7 +797,7 @@ CDentry* Server::rdlock_path_xlock_dentry(MDRequest *mdr, bool okexist, bool mus
   // -- lock --
   set<SimpleLock*> rdlocks, wrlocks, xlocks;
 
-  for (unsigned i=0; i<trace.size(); i++) 
+  for (int i=0; i<(int)trace.size(); i++) 
     rdlocks.insert(&trace[i]->lock);
   if (dn->is_null()) {
     xlocks.insert(&dn->lock);                 // new dn, xlock
@@ -1460,11 +1460,11 @@ void Server::handle_client_link(MDRequest *mdr)
   // create lock lists
   set<SimpleLock*> rdlocks, wrlocks, xlocks;
 
-  for (unsigned i=0; i<linktrace.size(); i++)
+  for (int i=0; i<(int)linktrace.size(); i++)
     rdlocks.insert(&linktrace[i]->lock);
   xlocks.insert(&dn->lock);
   wrlocks.insert(&dn->dir->inode->dirlock);
-  for (unsigned i=0; i<targettrace.size(); i++)
+  for (int i=0; i<(int)targettrace.size(); i++)
     rdlocks.insert(&targettrace[i]->lock);
   xlocks.insert(&targeti->linklock);
 
@@ -1712,7 +1712,7 @@ void Server::handle_client_unlink(MDRequest *mdr)
   // lock
   set<SimpleLock*> rdlocks, wrlocks, xlocks;
 
-  for (unsigned i=0; i<trace.size()-1; i++)
+  for (int i=0; i<(int)trace.size()-1; i++)
     rdlocks.insert(&trace[i]->lock);
   xlocks.insert(&dn->lock);
   wrlocks.insert(&dn->dir->inode->dirlock);
@@ -2105,13 +2105,13 @@ void Server::handle_client_rename(MDRequest *mdr)
   set<SimpleLock*> rdlocks, wrlocks, xlocks;
 
   // rdlock sourcedir path, xlock src dentry
-  for (unsigned i=0; i<srctrace.size()-1; i++) 
+  for (int i=0; i<(int)srctrace.size()-1; i++) 
     rdlocks.insert(&srctrace[i]->lock);
   xlocks.insert(&srcdn->lock);
   wrlocks.insert(&srcdn->dir->inode->dirlock);
 
   // rdlock destdir path, xlock dest dentry
-  for (unsigned i=0; i<desttrace.size(); i++)
+  for (int i=0; i<(int)desttrace.size(); i++)
     rdlocks.insert(&desttrace[i]->lock);
   xlocks.insert(&destdn->lock);
   wrlocks.insert(&destdn->dir->inode->dirlock);
